Learnt perceptron trick algorithm
In this perceptron trick algorithm, the basic idea is that it updates the weights for the inputs according to the update rule. 
First of all random points are selected, and then applied the WoXo + W1X1 + ...... + WnXn to get the prediction. 
If the difference between the actual class and the predicted (in terms of 1 and 0) then the update happens to the weights.
At last the code will run for n number of epochs to get a perfect classifying line.